Allow User to Change UID/GID Values
In the current version, the documentation advertises changing the UID and GID values in the `.env` file, however it doesn't appear it actually does anything when inspecting the codebase.

The current codebase allows you to set the UID and GID values when building the image, which isn't particularly useful when deploying containers.

My proposed changes allow for optionally changing the UID and GID values at runtime. To make this possible, the entire `Dockerfile` is ran as the `root` user. In the `docker-entrypoint.sh` file, which is ran at container runtime, if the UID or GID values are set in the environment variables, then they will be changed for the `anythingllm` user. The application is started as the `anythingllm` user through `gosu` rather than continuing as `root` user.

This proposed change is valuable to avoid permission issues on mapped volumes.